geolocation/geo-00152.html

<!DOCTYPE html>
<html>
  <head>
    <title>Geolocation Test: for latitude and longitude</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#geolocation">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#coordinates">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var t =async_test("Test checks that for latitude and longitude");
      setup({timeout: 20000});
      try {
        window.navigator.geolocation.getCurrentPosition(successCallback, unexpectedErrorCallback);
      } catch (err) {
        t.step(function () {
          assert_unreached("Error message: " + err.message);
        });
        t.done();
      }
      function successCallback (position) {
        t.step(function () {
          assert_true(position
            && typeof position.coords == 'object'
            && typeof position.coords.latitude == 'number'
            && typeof position.coords.longitude == 'number', "for latitude and longitude");
        });
        t.done();
      }
      function unexpectedErrorCallback (error) {
        t.step(function () {
          assert_unreached("Error message: " + error.message);
        });
        t.done();
      }
    </script>
  </body>
</html>
